refactor loadSourcText
