+++ src/actions/index.js
@@ -11,6 +11,7 @@ import * as ast from "./ast";
 import * as coverage from "./coverage";
 import * as projectTextSearch from "./project-text-search";
 import * as sourceSearch from "./source-search";
+import * as loadSourceText from "./sources/loadSourceText";
 
 export default Object.assign(
   {},
@@ -24,5 +25,6 @@ export default Object.assign(
   ast,
   coverage,
   projectTextSearch,
-  sourceSearch
+  sourceSearch,
+  loadSourceText
 );
